move unmerge_faces_tex into texture and split out materials
