+++ trimesh/visual/texture.py
@@ -3,7 +3,11 @@ import numpy as np
 import copy
 
 from . import color
+
 from .. import caching
+from .. import grouping
+
+from .material import SimpleMaterial
 
 
 class TextureVisuals(object):
@@ -168,97 +172,56 @@ class TextureVisuals(object):
         pass
 
 
-class Material(object):
-    def __init__(self, *args, **kwargs):
-        raise NotImplementedError('material must be subclassed!')
-
-
-class SimpleMaterial(Material):
+def unmerge_faces_tex(faces, faces_tex):
     """
-    Hold a single image texture.
+    Textured meshes can come with faces referencing vertex
+    indices (`v`) and an array the same shape which references
+    vertex texture indices (`vt`).
+
+    Parameters
+    -------------
+    faces : (n, d) int
+      References vertex indices
+    faces_tex : (n, d) int
+      References a list of UV coordinates
+
+    Returns
+    -------------
+    new_faces : (m, d) int
+      New faces for masked vertices
+    mask_v : (p,) int
+      A mask to apply to vertices
+    mask_vt : (p,) int
+      A mask to apply to vt array to get matching UV coordinates
     """
+    # stack into pairs of (vertex index, texture index)
+    stack = np.column_stack((faces.reshape(-1),
+                             faces_tex.reshape(-1)))
+    # find unique pairs: we're trying to avoid merging
+    # vertices that have the same position but different
+    # texture coordinates
+    unique, inverse = grouping.unique_rows(stack)
 
-    def __init__(self,
-                 image=None,
-                 diffuse=None,
-                 ambient=None,
-                 specular=None,
-                 **kwargs):
+    # only take the unique pairs
+    pairs = stack[unique]
+    # try to maintain original vertex order
+    order = pairs[:, 0].argsort()
+    # apply the order to the pairs
+    pairs = pairs[order]
 
-        # save image
-        self.image = image
+    # the mask for vertices, and mask for vt to generate uv coordinates
+    mask_v, mask_uv = pairs.T
 
-        # save material colors
-        self.ambient = ambient
-        self.diffuse = diffuse
-        self.specular = specular
+    # we re-ordered the vertices to try to maintain
+    # the original vertex order as much as possible
+    # so to reconstruct the faces we need to remap
+    remap = np.zeros(len(order), dtype=np.int64)
+    remap[order] = np.arange(len(order))
 
-        # save other keyword arguments
-        self.kwargs = kwargs
+    # the faces are just the inverse with the new order
+    new_faces = remap[inverse].reshape((-1, 3))
 
-    def to_color(self, uv):
-        return uv_to_color(uv, self.image)
-
-
-class PBRMaterial(Material):
-    """
-    Create a material for physically based rendering as
-    specified by GLTF 2.0:
-    https://git.io/fhkPZ
-
-    Parameters with `Texture` in them must be PIL.Image objects
-    """
-
-    def __init__(self,
-                 name=None,
-                 emissiveFactor=None,
-                 emissiveTexture=None,
-                 normalTexture=None,
-                 occlusionTexture=None,
-                 baseColorTexture=None,
-                 baseColorFactor=None,
-                 metallicFactor=None,
-                 roughnessFactor=None,
-                 metallicRoughnessTexture=None,
-                 doubleSided=False,
-                 alphaMode='OPAQUE',
-                 alphaCutoff=0.5):
-
-        # To to-float conversions
-        if baseColorFactor is not None:
-            baseColorFactor = np.array(baseColorFactor, dtype=np.float)
-        if emissiveFactor is not None:
-            emissiveFactor = np.array(emissiveFactor, dtype=np.float)
-
-        # (4,) float
-        self.baseColorFactor = color.to_rgba(baseColorFactor)
-
-        # (3,) float
-        self.emissiveFactor = color.to_rgba(emissiveFactor)
-
-        # float
-        self.metallicFactor = metallicFactor
-        self.roughnessFactor = roughnessFactor
-        self.alphaCutoff = alphaCutoff
-
-        # PIL image
-        self.normalTexture = normalTexture
-        self.emissiveTexture = emissiveTexture
-        self.occlusionTexture = occlusionTexture
-        self.baseColorTexture = baseColorTexture
-        self.metallicRoughnessTexture = metallicRoughnessTexture
-
-        # bool
-        self.doubleSided = doubleSided
-
-        # str
-        alphaMode = alphaMode
-
-    def to_color(self, uv):
-        color = uv_to_color(uv=uv, image=self.baseColorTexture)
-        if color is None and self.baseColorFactor is not None:
-            color = self.baseColorFactor.copy()
-        return color
+    return new_faces, mask_v, mask_uv
 
 
 def uv_to_color(uv, image):
